HandlerBox fixes for non-compliant strings
The handler is terminated by the string, so regardless as to whether
the string is properly nul terminated, we should be able to parse the
string greedily.  Add a string parser that will allow us to read bytes
greedily until the stream ends or a nul is encountered and then return
the resulting string.